refacto: Move route definitions into separate files.
